FrSky RSSI: split headers, cleanup, remove RssiPwmPeriod from UAVO def
